+++ front-end/src/Profile/EditPassword.css
@@ -1,0 +1,8 @@
+.input-border {
+    border-top-style: hidden;
+    border-right-style: hidden;
+    border-left-style: hidden;
+    border-bottom: 1px solid #ccc;
+    background-color: transparent;
+    margin-right: 10px;
+  }
