+++ front-end/src/signUp/ProfileImage.css
@@ -1,18 +1,54 @@
-.profile-image {
+.profile-image-container {
   display: flex;
   flex-direction: column;
   align-items: center;
   margin-bottom: 20px;
+}
+
+.profile-image {
+  position: relative;
+  width: 150px;
+  height: 150px;
   text-align: center;
+  cursor: pointer;
 }
 
 .profile-image img {
-  width: 150px;
-  height: 150px;
+  width: 100%;
+  height: 100%;
   object-fit: cover;
   border: 1px solid #ddd;
+  border-radius: 50%;
+  transition: all 0.3s;
 }
 
-.profile-image input[type='file'] {
-  margin-top: 10px;
+.profile-image:hover img {
+  filter: brightness(50%);
+}
+
+.profile-image-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0;
+  transition: opacity 0.3s;
+}
+
+.profile-image:hover .profile-image-overlay {
+  opacity: 1;
+}
+
+.profile-image-text {
+  color: #fff;
+  font-size: 14px;
+  font-weight: bold;
+}
+
+.profile-image-input {
+  display: none;
 }
